Cleaned up sprite drawing routines
Moved them into their own file (wl_scale.c) and made some improvements.

There is still a LOT of room for improvement (especially with the 3D sprites), but this should be a good start.

+++ id_pm.h
@@ -28,7 +28,7 @@ uint8_t *PM_GetEnd();
 
 byte *PM_GetTexture(int wallpic);
 
-uint16_t *PM_GetSprite(int shapenum);
+byte *PM_GetSprite(int shapenum);
 
 byte *PM_GetSound(int soundpagenum);
